requirements.txt
speedtest-cli
python-telegram
pickle
os
shutil
datetime
time
pathlib
